Update ParameterConsumer index type from short to int
Originally ParameterVisitable was built for ReusableMessage which
provides a "short getParameterCount()" method. Now that the
interface isn't bound to ReusableMessage it might as well use
integer.

+++ log4j-api/src/main/java/org/apache/logging/log4j/message/ReusableObjectMessage.java
@@ -114,7 +114,7 @@ public class ReusableObjectMessage implements ReusableMessage, ParameterVisitabl
 
     @Override
     public <S> void forEachParameter(ParameterConsumer<S> action, S state) {
-        action.accept(obj, (short) 0, state);
+        action.accept(obj, 0, state);
     }
 
     @Override
